feat: cryptocurrency services

+++ TradeBot/src/main/java/com/domingueti/tradebot/modules/Cryptocurrency/services/DeleteCryptocurrencyByIdService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.domingueti.tradebot.exceptions.NotFoundException;
+import com.domingueti.tradebot.modules.Cryptocurrency.models.Cryptocurrency;
 import com.domingueti.tradebot.modules.Cryptocurrency.repositories.CryptocurrencyRepository;
 
 @Service
@@ -14,7 +16,15 @@ public class DeleteCryptocurrencyByIdService {
 	
 	@Transactional
 	public void execute(Long id) {
-//		validator.execute(id); 
+		
+		Cryptocurrency cryptocurrency = cryptocurrencyRepository.findByIdAndDeletedAtIsNull(id);
+		
+//		validator.execute(cryptocurrency);
+		
+		if (cryptocurrency == null) {
+			throw new NotFoundException("Cryptocurrency not found with given ID: " + id);
+		}
+
 		cryptocurrencyRepository.deleteById(id);
 	}
